+++ sql-infer-cli/src/codegen/sqlalchemy_v2.rs
@@ -47,10 +47,10 @@ fn to_py_input_type(item: &QueryItem) -> String {
         | SqlType::Text
         | SqlType::Json
         | SqlType::Jsonb => "str",
-
         SqlType::Float4 | SqlType::Float8 => "float",
         SqlType::Interval => "timedelta",
         SqlType::Bit { .. } | SqlType::VarBit { .. } => "str",
+        SqlType::Unknown => "Any",
     }
     .to_owned();
     match item.nullable {
@@ -82,6 +82,7 @@ fn to_pydantic_input_type(item: &QueryItem) -> String {
         SqlType::Float4 | SqlType::Float8 => "float",
         SqlType::Interval => "timedelta",
         SqlType::Bit { .. } | SqlType::VarBit { .. } => "str",
+        SqlType::Unknown => "Any",
     }
     .to_owned();
     match item.nullable {