+++ sql-infer-cli/src/main.rs
@@ -1,6 +1,7 @@
 pub mod codegen;
 mod commands;
 pub mod config;
+pub mod schema;
 pub mod utils;
 
 use clap::*;
@@ -8,13 +9,14 @@ use commands::Generate;
 use tracing::Level;
 use tracing_subscriber::FmtSubscriber;
 
-use crate::commands::analyze::Analyze;
+use crate::commands::{analyze::Analyze, schema::Schema};
 
 #[derive(Parser)]
 #[command(name = "sql-infer", bin_name = "sql-infer")]
 enum Command {
     Generate(Generate),
     Analyze(Analyze),
+    Schema(Schema),
 }
 
 #[tokio::main]
@@ -29,6 +31,7 @@ async fn main() -> Result<(), String> {
     let res = match command {
         Command::Generate(args) => args.run().await,
         Command::Analyze(analyze) => analyze.run().await,
+        Command::Schema(schema) => schema.run().await,
     };
     if let Err(err) = res {
         return Err(err.to_string());